feat. 피드 좋아요, 피드 좋아요순 정렬, 피드 좋아요 취소
기능 추가 구현
테스트까지 완료

+++ src/feeds/feeds.service.ts
@@ -40,10 +40,10 @@ export class FeedsService {
     });
     const createdFeedId = createdFeed.identifiers[0].id;
     const favoriteConfirm = await this.favoriteRepository.findOne({
-      where: {id: favorite_id, deletedAt:null}
-    })
-    if(!favoriteConfirm) {
-      return {errorMessage: '이미 삭제된 favorite_id입니다.'}
+      where: { id: favorite_id, deletedAt: null },
+    });
+    if (!favoriteConfirm) {
+      return { errorMessage: '이미 삭제된 favorite_id입니다.' };
     }
     const deleteFavoriteCart = await this.favoriteRepository
       .createQueryBuilder()
@@ -58,18 +58,35 @@ export class FeedsService {
   }
 
   async getFeeds() {
-    // 최신순, 좋아요순 생각해야됨(디폴트는 좋아요순)
-    // 쿼리문으로 생각해보기 (order by, group by)
-    const allFeeds = await this.feedRepository.find({
-      where: { deletedAt: null },
-      select: { title: true },
-    });
+    // const allFeeds = await this.feedLikeRepository
+    //   .createQueryBuilder('fl')
+    //   .select('f.id', 'id')
+    //   .addSelect('f.title', 'title')
+    //   .addSelect('COUNT(*)', '피드 좋아요수')
+    //   .innerJoin('fl.feed', 'f')
+    //   .groupBy('fl.feed_id')
+    //   .orderBy('피드 좋아요수', 'DESC')
+    //   .getRawMany();
+    const allFeeds = await this.feedLikeRepository
+      .query(`SELECT f.id, f.title, COUNT(*) as '피드좋아요'
+      FROM feed_like fl INNER JOIN feed f ON f.id = fl.feed_id
+      GROUP BY fl.feed_id ORDER BY '피드좋아요' DESC
+      `);
     return allFeeds;
   }
 
   async getFeed(id: number) {
     const findFeed = await this.feedRepository.findOne({
       where: { id },
+      relations: ['users'],
+      select: [
+        'id',
+        'user_id',
+        'title',
+        'description',
+        'createdAt',
+        'updatedAt',
+      ],
     });
     if (!findFeed) {
       return { errorMessage: '잘못된 id입니다.' };
@@ -77,34 +94,50 @@ export class FeedsService {
     return findFeed;
   }
 
-  async updateFeed(id: number, title: string, description: string) {
-    await this.feedRepository.update(id, { title, description });
-  }
-  async deleteFeed(id: number) {
+  async updateFeed(
+    id: number,
+    user_id: number,
+    title: string,
+    description: string,
+  ) {
     const findFeed = await this.getFeed(id);
     if (_.isNil(findFeed)) {
       throw new NotFoundException(
         `피드번호 ${id}번의 피드를 찾을 수 없습니다.`,
       );
     }
-    // if(findFeed.user_id !== user_id) {
-    //   throw new UnauthorizedException('작성자만 삭제 가능합니다.')
-    // }
+    if (findFeed['user_id'] !== user_id) {
+      throw new UnauthorizedException('작성자만 수정 가능합니다.');
+    }
+    await this.feedRepository.update(id, { title, description });
+  }
+  async deleteFeed(id: number, user_id: number) {
+    const findFeed = await this.getFeed(id);
+    if (_.isNil(findFeed)) {
+      throw new NotFoundException(
+        `피드번호 ${id}번의 피드를 찾을 수 없습니다.`,
+      );
+    }
+    if (findFeed['user_id'] !== user_id) {
+      throw new UnauthorizedException('작성자만 삭제 가능합니다.');
+    }
     await this.feedRepository.softDelete(id);
     return { message: `피드번호 ${id}번의 피드가 삭제되었습니다.` };
   }
-  async getFeedLikes(id: number) {
-    // 피드 좋아요 개수
-    await this.feedLikeRepository.query(`
-    SELECT COUNT(*) FROM feedlike WHERE feed_id = ${id}
-    `);
+  async getFeedLikes(feed_id: number) {
+    const feedLikeCount = await this.feedLikeRepository
+      .createQueryBuilder('feed_like')
+      .select('COUNT(*)', 'count')
+      .where('feed_like.feed_id = :feed_id', { feed_id })
+      .getRawOne();
+    return feedLikeCount;
   }
-  async feedLike(feed_id: number) {
-    // user_id 넣어야됨
-    await this.feedLikeRepository.insert({ feed_id });
+  async feedLike(feed_id: number, user_id: number) {
+    await this.feedLikeRepository.insert({ feed_id, user_id });
+    return { message: '좋아요 추가' };
   }
-  async feedLikeCancel(feed_id: number) {
-    // user_id 넣어야됨
-    await this.feedLikeRepository.delete({ feed_id });
+  async feedLikeCancel(feed_id: number, user_id: number) {
+    await this.feedLikeRepository.delete({ feed_id, user_id });
+    return { message: '좋아요 취소' };
   }
 }
